+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## vNEXT
 - Add slither suggestions. (#26)
+- Drain expired vouchers and withdraw funds. (#25)
 - Add slither github action. (#24)
 - Top up voucher. (#23)
 - Claim task part 2 - Add voucher tests. (#21)